Created a new PlayerPrefab called "Player - Test".  See Dungeon scene for Example.  Player can pick up objects and are immediately equipped.  Pressing "Right Arrow" changes currently held item.  Health potions are picked up automatically.
Known Problems:
Player can only pick up two items, it refuses to drop an item if full.  Cannot find a fix.
Health items are not restoring health on use.  They are deincrimenting on use, but no health change.

NOTES:
Weapon drop prefabs can be assembled quickly using the current format for the 3 current weapons on use.  But fixing the problems above will have to come first, or more weapons won't matter if we are unable to swap them when picking up.